Further small improvements
-EN readme
- option for CLAHE only
