+++ requirements.txt
@@ -1,5 +1,8 @@
-opencv_python==4.4.0.46
+python 3
+opencv-python==4.4.0.46
 numpy==1.19.2
 astropy==4.2
 PySimpleGUI==4.38.0
+scipy
+matplotlib
