+++ windows_setup.bat
